requirements.txt
accelerate==0.34.2
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.5.0
attrs==24.2.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
dbus-python==1.2.16
diffusers==0.30.3
distro==1.4.0
distro-info==0.23+ubuntu1.1
exceptiongroup==1.2.2
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gradio==4.44.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
-e git+https://github.com/huggingface/huggingface_hub.git@64bcff5c7725f29514b15d4db3a7a92e3955e577#egg=huggingface_hub
humanfriendly==10.0
idna==2.8
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
peft==0.12.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.2
psutil==5.9.8
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.36.0
PyMatting==1.1.12
pyparsing==3.1.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.6
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.14.0
sniffio==1.3.1
spaces==0.30.2
starlette==0.38.5
sympy==1.13.3
tifffile==2024.8.30
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.30.6
websockets==12.0
zipp==3.20.2